refactor(web): extract app display components
