+++ apps/web/src/lib/text.ts
@@ -1,0 +1,5 @@
+export const normalizeSuggestion = (value?: string): string | undefined => {
+  if (!value) return undefined;
+  const trimmed = value.trim();
+  return trimmed.length > 0 ? trimmed : undefined;
+};
